Add autogen.sh and update README.md
A command or script to generate configure file is needed. We encapsulated into autogen.sh. Note that we didn't handle no libtoolize on macOS due to we probably have to seprarate sub-command of autoreconf (making alias didn't work as tested). Better to let autoreconf --install handles this for us.

So in case of no libtoolize on macOS when execute, you will see error message. Try to symlink glibtoolize to libtoolize to satisfy the requirement.